% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exercise_2_3_1}
\alias{exercise_2_3_1}
\title{Run example 2.3.1}
\usage{
exercise_2_3_1(MAXTIME = 100L, PRNTTIME = 10L)
}
\arguments{
\item{MAXTIME}{integer}

\item{PRNTTIME}{integer}
}
\description{
Run example 2.3.1
}
