% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exercise_2_6_1}
\alias{exercise_2_6_1}
\title{Run example 2.6.1}
\usage{
exercise_2_6_1(MAXTIME = 100, PRNTTIME = 10L, NUMRUNS = 1)
}
\arguments{
\item{MAXTIME}{double}

\item{PRNTTIME}{int}

\item{NUMRUNS}{double}
}
\description{
Run example 2.6.1
}
